Refactor admin users routes into dedicated module
- Extract admin user routes from http.ts into server/src/routes/admin/users.ts
- Move GET /api/admin/users (unified user list)
- Move GET /api/admin/users/memberships (export)
- Move POST /api/admin/users/sync-workos (backfill from WorkOS)
- Update admin-users.html to use new sync endpoint path
- Clean up http.ts imports

This refactor improves code organization by keeping admin routes
in their own modules under routes/admin/.
