+++ server/src/routes/admin/index.ts
@@ -8,6 +8,7 @@ export { createAdminSlackRouter } from './slack.js';
 export { createAdminEmailRouter } from './email.js';
 export { createAdminFeedsRouter } from './feeds.js';
 export { createAdminNotificationChannelsRouter } from './notification-channels.js';
+export { createAdminUsersRouter } from './users.js';
 
 // Core admin route setup functions
 export { setupProspectRoutes } from './prospects.js';
